--- matiushin_medvedev/__init__.py
@@ -1,17 +1,0 @@
-from flask import Flask
-
-from matiushin_medvedev.db_config import db
-from matiushin_medvedev.web.educational_plan_controller import educational_plan_controller
-from matiushin_medvedev.web.gradebook_controller import gradebook_controller
-from matiushin_medvedev.web.student_controller import student_controller
-
-app = Flask(__name__)
-app.register_blueprint(student_controller)
-app.register_blueprint(educational_plan_controller)
-app.register_blueprint(gradebook_controller)
-app.config['SQLALCHEMY_DATABASE_URI'] = "postgresql://education:password@localhost:5432/education"
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db.init_app(app)
-
-if __name__ == 'main':
-    app.run()
