+++ matiushin_medvedev/backend/web/gradebook_controller.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request
 from flask_api import status
 
-from matiushin_medvedev.web.controller_utils import convert_to_json
+from matiushin_medvedev.backend.web.controller_utils import convert_to_json
 
 gradebook_controller = Blueprint('gradebook_controller', __name__)
